Split component artifact CLI imports
